added working ddl for view switching
